// src/utils/Types.tsx
enum ActionTypes {
    AUTH_LOADING = "AUTH_LOADING",
    AUTH_AUTHENTICATION = "AUTH_AUTHENTICATION",
    AUTH_VALIDATE = "AUTH_VALIDATE",
    AUTH_RESET = "AUTH_RESET",
    AUTH_VALIDATE_EMAIL = "AUTH_VALIDATE_EMAIL",
    AUTH_FAIL = "AUTH_FAIL",
    AUTH_LOGOUT = "AUTH_LOGOUT",
    AUTH_ACTIVITY = "AUTH_ACTIVITY",
    AUTH_PASSWORD_RESULT = "AUTH_PASSWORD_RESULT",
    AUTH_ACTIVITY_RESET = "AUTH_ACTIVITY_RESET",
};
export default ActionTypes;